Begin breaking out window boundary by type
